Refactor: Constants moved to config folder
