+++ backend/src/config/index.ts
@@ -1,2 +1,3 @@
 export * from './App';
+export * from './Constants';
 export * from './Envs';
